+++ CHANGELOG.md
@@ -87,6 +87,7 @@ BUG FIXES:
 ## 0.65.0 (October 14, 2021)
 FEATURES:
 * mdb: support Schema Registry in `yandex_mdb_kafka_cluster`
+* resourcemanager: data source `yandex_resourcemanager_cloud` now provides `folders` attribute
 
 FEATURES:
 * **New Resource:** `yandex_kms_symmetric_key_iam_binding`